[graphite] Use 32 bit integers for SSBO indices
This hasn't been a problem in the past, because we've limited
the number of draws in a DrawPass to at most 16k (defined in
DrawList::kMaxRenderSteps), and we batch all uniform uploads for a
single pipeline such that all storage buffer indices will be within
16k elements of their respective storage buffer binding offset.

With changes to uniform management such that we no longer batch uniform
uploads by pipeline in http://review.skia.org/898257, we always bind to
the start of storage buffers and index relative to that binding. So our
indices may be greater than 16k, maybe even more than 64k, and we need
more than 16 bits to represent them.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/904783
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>

+++ src/gpu/graphite/Renderer.h
@@ -67,7 +67,7 @@ public:
     // The DrawWriter is configured with the vertex and instance strides of the RenderStep, and its
     // primitive type. The recorded draws will be executed with a graphics pipeline compatible with
     // this RenderStep.
-    virtual void writeVertices(DrawWriter*, const DrawParams&, skvx::ushort2 ssboIndices) const = 0;
+    virtual void writeVertices(DrawWriter*, const DrawParams&, skvx::uint2 ssboIndices) const = 0;
 
     // Write out the uniform values (aligned for the layout), textures, and samplers. The uniform
     // values will be de-duplicated across all draws using the RenderStep before uploading to the
