+++ src/gpu/graphite/render/PerEdgeAAQuadRenderStep.h
@@ -23,7 +23,7 @@ public:
     std::string vertexSkSL() const override;
     const char* fragmentCoverageSkSL() const override;
 
-    void writeVertices(DrawWriter*, const DrawParams&, skvx::ushort2 ssboIndices) const override;
+    void writeVertices(DrawWriter*, const DrawParams&, skvx::uint2 ssboIndices) const override;
     void writeUniformsAndTextures(const DrawParams&, PipelineDataGatherer*) const override;
 
 private:
